refactor(web): extract app shell header and footer components
